requirements.txt
cmake==3.22.6
IPython==7.34.0
imutils==0.5.4
numpy==1.21.6
pyqt5==5.15.4
pyqt5-tools==5.15.4.3.2
PyQt-builder==1.13.0
mysql-connector-python==8.0.30
pillow==7.1.2
wheel==0.34.2
opencv-python==4.5.5.64
deepface==0.0.75
dlib==19.24.0
face_recognition==1.3.0
pyserial==3.5